+++ app/models/planned_disbursement.rb
@@ -15,7 +15,9 @@ class PlannedDisbursement < ApplicationRecord
     :providing_organisation_name,
     :providing_organisation_type,
     :receiving_organisation_name,
-    :receiving_organisation_type
+    :receiving_organisation_type,
+    :financial_quarter,
+    :financial_year
   validates :value, inclusion: {in: 0.01..99_999_999_999.00}
   validates_with EndDateNotBeforeStartDateValidator, if: -> { period_start_date.present? && period_end_date.present? }
   validates :period_start_date, :period_end_date, date_within_boundaries: true