+++ spec/services/update_planned_disbursement_spec.rb
@@ -12,6 +12,26 @@ RSpec .describe UpdatePlannedDisbursement do
       expect(result.object.receiving_organisation_name).to eq("An Organisation")
     end
 
+    context "when financial quarter and year are provided" do
+      it "sets the period start and end dates" do
+        financial_quarter = "1"
+        financial_year = "2020"
+        result = described_class.new(planned_disbursement: planned_disbursement).call(attributes: {financial_quarter: financial_quarter, financial_year: financial_year})
+        expect(result.object.period_start_date).to eq "2020-04-01".to_date
+        expect(result.object.period_end_date).to eq "2020-06-30".to_date
+      end
+    end
+
+    context "when start and end date are provided" do
+      it "sets the financial quarter and year" do
+        start_date = "1 April 2020"
+        end_date = "30 June 2020"
+        result = described_class.new(planned_disbursement: planned_disbursement).call(attributes: {period_start_date: start_date, period_end_date: end_date})
+        expect(result.object.financial_quarter).to eq 1
+        expect(result.object.financial_year).to eq 2020
+      end
+    end
+
     it "returns a Result with the success set to true" do
       allow(planned_disbursement).to receive(:valid?).and_return(true)
       allow(planned_disbursement).to receive(:save!).and_return(true)